// src/Web/Desnz.Chmm.Web/ClientApp/src/app/stores/scheme-year-configuration/dtos/scheme-year-configuration.dto.ts
export interface SchemeYearConfigurationDto {
  percentageCap: number;
  targetMultiplier: number;
  gasCreditsCap: number;
  oilCreditsCap: number;
  gasBoilerSalesThreshold: number;
  oilBoilerSalesThreshold: number;
  targetRate: number;
  creditCarryOverPercentage: number;
  alternativeRenewableSystemFuelTypeWeightingValue: number;
  alternativeFossilFuelSystemFuelTypeWeightingValue: number;
}